component files should have the same name as their default functions
